Small fixes

+++ src/test/groovy/me/xdrop/jrand/generators/basics/FloatGeneratorTest.groovy
@@ -4,11 +4,11 @@ class FloatGeneratorTest extends GroovyTestCase {
     def instance = {-> new FloatGenerator()}
 
     void testMax() {
-        assertTrue instance().max(2.0).gen() <= 2
+        assertTrue instance().max(2.0 as float).gen() <= 2
     }
 
     void testMin() {
-        assertTrue instance().min(5.0).gen() >= 5
+        assertTrue instance().min(5.0 as float).gen() >= 5
     }
 
     void testRange() {
